Refactored firebase setup outside FirestoreSingleton
